terrain ids now working, setting type for one terrain quad
All fragments that belong to a quad get their terrain id from its front left vertex/sample point. This is the vertex with the minimum xz coordinates, or minimun row/column number in physx.

For now, the two triangles in a quad use the same terrain id.
 - in physx; materialId0==materialId1
 - in osg/opengl: terrainTypeIDs texture
The orientation of the diagonal in a quad in osg depends on the curvature of its normals, so we would need to set the physx diagonal orientation to the osg orientation, what is not done in the moment.